app/src/requirements.txt
pymongo
uvicorn
fastapi
pycryptodome
